:truck: move a component I missed into the right place
